# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EcoTrack API",
        "description": "Système de traçabilité intelligente pour la Supply Chain (Mobile & IoT).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8003",
            "description": "Serveur Local"
        }
    ],
    "paths": {
...
        "/api/worker/submit-scan": {
            "post": {
                "tags": [
                    "Simulation Ouvrier"
                ],
                "summary": "Simuler la soumission d'un scan par un ouvrier",
                "description": "Route de \"Passerelle\" pour simuler l'envoi depuis une App Mobile",
                "operationId": "56b8c5deb8090617ef2dafa953d04fee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_uuid",
                                    "status"
                                ],
                                "properties": {
                                    "product_uuid": {
                                        "type": "string",
                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "processed",
                                            "in_transit",
                                            "delivered"
                                        ]
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "example": 6.36
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "example": 2.43
                                    },
                                    "location_name": {
                                        "type": "string",
                                        "example": "Entrepôt de Tri A"
                                    },
                                    "metadata": {
                                        "type": "object",
                                        "example": {
                                            "condition": "good",
                                            "battery": "85%"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Scan métier simulé avec succès"
                    },
                    "422": {
                        "description": "Données invalides ou UUID inexistant"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez votre token reçu lors du login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}